minishell

+++ execc/unset.c
@@ -6,44 +6,51 @@
 /*   By: ckannane <ckannane@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/02 10:05:00 by ckannane          #+#    #+#             */
-/*   Updated: 2023/09/11 15:54:01 by ckannane         ###   ########.fr       */
+/*   Updated: 2023/09/22 20:34:48 by ckannane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
-void delete_val(t_val **head, char *str)
+void	delete_val(t_val **head, char *str)
 {
-	t_val *current = *head;
-	t_val *pre = NULL;
+	t_val	*current;
+	t_val	*pre;
 
-	if(current != NULL && ft_strcmp(current->name,str) == 0)
+	current = *head;
+	if (current != NULL && ft_strcmp(current->name, str) == 0)
 	{
 		*head = current->next;
-        free(current);
-        return;
+		free(current->value);
+		free(current->name);
+		free(current);
+		return ;
 	}
-	while (current != NULL && ft_strcmp(current->name,str) != 0) {
-        pre = current;
-        current = current->next;
-    }
-	if (current == NULL) {
-		printf("nor found\n");
-        return;
-    }
-	printf("we found him\n");
+	while (current != NULL && ft_strcmp(current->name, str) != 0)
+	{
+		pre = current;
+		current = current->next;
+	}
+	if (current == NULL)
+		return ;
 	pre->next = current->next;
+	free(current->value);
+	free(current->name);
 	free(current);
 }
 
 void	ft_unset(t_com *com, t_zid *zone)
 {
-	int i = 0;
-	while(com->arg[i])
+	int	i;
+
+	i = 0;
+	while (com->arg[i])
 	{
-		printf("the unset applyed on (%s)\n",com->arg[i]);
-		delete_val(&zone->env,com->arg[i]);
-		delete_val(&zone->exp,com->arg[i]);
+		if (ft_strcmp(com->arg[i], "_") && (zone->env || zone->exp))
+		{
+			delete_val(&zone->exp, com->arg[i]);
+			delete_val(&zone->env, com->arg[i]);
+		}
 		i++;
 	}
 }
